Add tokio support
This was pretty straight forward to port! Changes:

* async functionality for all the functions that are used for comms
* small type changes throughout to match the tokio-i3ipc version

There is a small hitch in window marks - they are not currently visible
in tokio-i3ipc. I opened this pr: https://github.com/leshow/tokio-i3ipc/pull/43
to fix that. In the mean time the Cargo.toml points at my branch.

+++ src/collapse.rs
@@ -2,13 +2,13 @@ extern crate log;
 
 use std::collections::HashSet;
 
-use i3ipc::reply::Node;
+use tokio_i3ipc::reply::Node;
 
 use crate::ext::NodeSearch;
 
-fn find_candidate(root: &Node) -> Vec<(&Node, i64)> {
+fn find_candidate(root: &Node) -> Vec<(&Node, usize)> {
     let mut leaves_seen = HashSet::new();
-    let mut res: Vec<(&Node, i64)> = Vec::with_capacity(2);
+    let mut res: Vec<(&Node, usize)> = Vec::with_capacity(2);
     for s in root.preorder() {
         println!("Walk to: id({})", s.n.id);
         // skip root since it's a workspace and that gets messy
